fix: Temporarily disable events subscription in useBackgroundBashCount
The Lens client doesn't yet support calling .subscribe() on query endpoints
that have a subscription handler. This is a known limitation.

For now, the hook relies on initial polling which is sufficient.

TODO: Update Lens client to expose .subscribe() method on query endpoints
with subscription handlers (dual query/subscribe endpoints).